requirements.txt
allauth==0.0.1
asgiref==3.9.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
